+++ docs/CHANGELOG.rst
@@ -22,6 +22,8 @@ Added
 
 Changed
 -------
+- **BACKWARD INCOMPATIBLE:** Rewrite ``differentialexpression-deseq2``
+  to Python
 - Add format parameter to ``macs2-callpeak``